requirements.txt
Django==2.2
django-crispy-forms>=1.0,<2.0
django-filter>=2.1.0,<2.2.0
django-guardian>=1.0,<2.0
djangorestframework>=3.9.0,<4.0
djangorestframework-jwt>=1.11.0,<1.12.0
psycopg2>=2.7,<3.0
requests>=2.0,<3.0
requests-toolbelt>0.9.0,<1.0.0
markdown>=3.0,<4.0
PyLD>=1.0,<2.0
serpy>=0.3.0,<0.4.0
pika>=1.0.0,<2.0.0
django-notifs>=2.6.0,<2.7.0
ipython>=7.0.0,<8.0.0
django_extensions>=2.0.0,<3.0.0
Werkzeug>=0.15.0,<0.16.0
watchdog>=0.9.0,<0.10.0
itsdangerous>=1.1.0,<2.0.0
MarkupSafe>=1.1.0,<2.0.0
django-allauth>=0.39.1,<0.40.0
Pillow>=6.0.0,<7.0.0
django-grappelli>=2.12.2,<3.0.0
django-filebrowser>=3.11.1,<4.0.0
mysqlclient
pytest
pylint
